fix deregisterValidator tests

+++ l1-contracts/test/helpers/UnitTestHelper.sol
@@ -37,7 +37,7 @@ contract UnitTestHelper is Test, BaseScript {
     address public operator = vm.addr(operatorPrivateKey);
     address public podOwner = makeAddr("podOwner");
 
-    uint256 public constant DEREGISTRATION_DELAY = 65;
+    uint64 public constant DEREGISTRATION_DELAY = 65;
 
     modifier fuzzedAddress(address addr) virtual {
         vm.assume(fuzzedAddressMapping[addr] == false);
